protège la page d'accès vers les indicateurs vsme accessible uniquement aux collaborateurs d'une entreprise qualifiée et ajoute une url générique
